requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.1
djangorestframework==3.11.1
folium==0.10.1
idna==2.10
isort==5.5.2
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.2
py4j==0.10.9
pylint==2.6.0
pyparsing==2.4.7
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scipy==1.5.2
seaborn==0.9.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
mysqlclient==2.0.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.5.0